+++ simple_value_object/decorators.py
@@ -1,5 +1,6 @@
 def invariant(func):
-    def invariant_func_wrapper(cls, instance):
-        return func(cls, instance)
+    def invariant_func_wrapper(instance):
+        return func(instance)
 
+    invariant_func_wrapper.name = func.__name__
     return invariant_func_wrapper